+++ Sources/Vaux/Builders.swift
@@ -7,16 +7,6 @@
 
 import Foundation
 
-/// Heading weight for the heading tag.
-public enum HeadingWeight {
-    case h1
-    case h2
-    case h3
-    case h4
-    case h5
-    case h6
-}
-
 /// Inserts the top level `<html>` element into the HTML document, and closes with `</html>` after the contents of the closure.
 /// - Parameters:
 ///   - child: `HTML` content to go inside the `<html></html>` element.
@@ -62,7 +52,7 @@ public func lineBreak() -> HTML {
 ///   - weight: The weight of the heading, such as `<h1>` or `<h4>`.
 ///   - child: `HTML` content to go inside the `<h1></h1>` element.
 public func heading(_ weight: HeadingWeight, @HTMLBuilder child: () -> HTML) -> HTML {
-  return HTMLNode(tag: "\(weight)", child: child())
+  return HTMLNode(tag: weight.rawValue, child: child())
 }
 
 /// Inserts a `<p>` element into the HTML document, and closes with `</p>` after the contents of the closure.
@@ -164,20 +154,6 @@ public func custom(tag: String, @HTMLBuilder child: () -> HTML) -> HTML {
 /// This allows you to iterate over a collection of data that will be rendered into HTML. By inspecting the variable passed into the closure, you can choose to insert it via a HTML node, or do nothing.
 public func forEach<Coll: Collection>(_ data: Coll, @HTMLBuilder content: @escaping (Coll.Element) -> HTML) -> HTML {
   return MultiNode(children: data.map(content))
-}
-
-public enum Scope: String {
-  case row = "row"
-  case column = "column"
-  case rowGroup = "rowgroup"
-  case columnGroup = "colgroup"
-}
-
-public enum Alignment: String {
-  case left = "left"
-  case right = "right"
-  case center = "center"
-  case justified = "justified"
 }
 
 extension HTML {
